% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrollment.R
\name{Enrollment}
\alias{Enrollment}
\title{A Cat Function}
\usage{
Enrollment(myschoolyear = 2016, currentDate = paste(substring(Sys.Date(), 6,
  7), substring(Sys.Date(), 9, 10), substring(Sys.Date(), 1, 4), sep = "/"),
  maxDate = "12/31/9999", PrimaryOnly = "N", EnrolledOn = NULL,
  includeFRL = "N", includeSPED = "N", includeELL = "N",
  includeRace = "N", includeTitleITargeted = "N",
  includeTitleISchool = "N", includeLAP = "N")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

