% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrollment.R
\name{Enrollment}
\alias{Enrollment}
\title{Student Enrollment}
\usage{
Enrollment(myschoolyear = 2016, currentDate = paste(substring(Sys.Date(), 6,
  7), substring(Sys.Date(), 9, 10), substring(Sys.Date(), 1, 4), sep = "/"),
  maxDate = "12/31/9999", PrimaryOnly = "N", EnrolledOn = NULL,
  includeFRL = "N", includeSPED = "N", includeELL = "N",
  includeRace = "N", includeTitleITargeted = "N",
  includeTitleISchool = "N", includeLAP = "N")
  }

{\arguments{\item{myschoolyear:}{School Year to run}}
 \arguments{\item{currentDate:}{Defaults to todays date.  Can run any 'as of'}}
 \arguments{\item{maxDate:}{ Max Date - defaults to 12/31/999}}
 \arguments(\item{PrimaryOnly:}{ Include only primary school enrollment ('Y') or all Enrollment('N')}}}
            
\description{
This function will pull all students enrolled at any point during the school year.
}
\examples{
2016 Enrollment including FRL - Enrollment(myschoolyear=2016,includeFRL='Y')
}
\keyword{cats}

